[REFACTOR]hooks dir 위치 수정
